#106 - 비즈니스 로직 groupService -> groupLikeService 로 이동
