+++ src/main/java/com/ting/ting/configuration/AppConfig.java
@@ -29,7 +29,7 @@ public class AppConfig {
 
     @Bean
     public GroupLikeService groupLikeService() {
-        return new GroupLikeServiceImpl(userRepository, groupRepository, groupMemberRepository, groupLikeToJoinRepository, groupLikeToDateRepository);
+        return new GroupLikeServiceImpl(userRepository, groupRepository, groupMemberRepository, groupDateRequestRepository, groupLikeToJoinRepository, groupLikeToDateRepository);
     }
 
     @Bean
