+++ src/main/java/com/ting/ting/service/GroupLikeService.java
@@ -1,6 +1,15 @@
 package com.ting.ting.service;
 
+import com.ting.ting.dto.response.DateableGroupResponse;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
 public interface GroupLikeService {
+
+    /**
+     * 팀 기준 - 찜한 목록 조회
+     */
+    Page<DateableGroupResponse> findGroupLikeToDateList(Long groupId, Long userId, Pageable pageable);
 
     /**
      * 같은 성별의 팀 찜하기
