+++ src/main/java/com/ting/ting/service/GroupLikeServiceImpl.java
@@ -1,11 +1,23 @@
 package com.ting.ting.service;
 
 import com.ting.ting.domain.*;
+import com.ting.ting.domain.constant.LikeStatus;
+import com.ting.ting.domain.constant.RequestStatus;
+import com.ting.ting.domain.custom.GroupIdWithLikeCount;
+import com.ting.ting.dto.response.DateableGroupResponse;
 import com.ting.ting.exception.ErrorCode;
 import com.ting.ting.exception.ServiceType;
 import com.ting.ting.repository.*;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Transactional
 @Component
@@ -14,17 +26,61 @@ public class GroupLikeServiceImpl extends AbstractService implements GroupLikeSe
     private final UserRepository userRepository;
     private final GroupRepository groupRepository;
     private final GroupMemberRepository groupMemberRepository;
+    private final GroupDateRequestRepository groupDateRequestRepository;
     private final GroupLikeToJoinRepository groupLikeToJoinRepository;
     private final GroupLikeToDateRepository groupLikeToDateRepository;
 
-    public GroupLikeServiceImpl(UserRepository userRepository, GroupRepository groupRepository, GroupMemberRepository groupMemberRepository, GroupLikeToJoinRepository groupLikeToJoinRepository,
-                                GroupLikeToDateRepository groupLikeToDateRepository){
+    public GroupLikeServiceImpl(UserRepository userRepository, GroupRepository groupRepository, GroupMemberRepository groupMemberRepository, GroupDateRequestRepository groupDateRequestRepository,
+                                GroupLikeToJoinRepository groupLikeToJoinRepository, GroupLikeToDateRepository groupLikeToDateRepository){
         super(ServiceType.GROUP_MEETING);
         this.userRepository = userRepository;
         this.groupRepository = groupRepository;
         this.groupMemberRepository = groupMemberRepository;
+        this.groupDateRequestRepository = groupDateRequestRepository;
         this.groupLikeToJoinRepository = groupLikeToJoinRepository;
         this.groupLikeToDateRepository = groupLikeToDateRepository;
+    }
+
+    @Override
+    public Page<DateableGroupResponse> findGroupLikeToDateList(Long groupId, Long userId, Pageable pageable) {
+        Group group = loadGroupByGroupId(groupId);
+        User member = loadUserByUserId(userId);
+
+        GroupMember memberRecordOfUser = groupMemberRepository.findByGroupAndMember(group, member).orElseThrow(() ->
+                throwException(ErrorCode.REQUEST_NOT_FOUND, String.format("User(id: %d) is not a member of the Group(id: %d)", userId, group))
+        );
+
+        Page<GroupIdWithLikeCount> idAndLikeCountOfGroupsLikeToDate = groupLikeToDateRepository.findAllToGroupIdAndLikeCountByFromGroupMember_Group(group, pageable);
+        Map<Long, Integer> groupIdWithLikeCountMap = idAndLikeCountOfGroupsLikeToDate.stream().collect(Collectors.toMap(GroupIdWithLikeCount::getGroupId, GroupIdWithLikeCount::getLikeCount));
+
+        // 찜한 기록이 없는 경우는 바로 return
+        if (idAndLikeCountOfGroupsLikeToDate.isEmpty()) {
+            return new PageImpl<>(List.of(), pageable, idAndLikeCountOfGroupsLikeToDate.getTotalElements());
+        }
+
+        Set<Group> likedGroups = groupRepository.findAllWithMembersInfoByIdIn(idAndLikeCountOfGroupsLikeToDate.stream().map(GroupIdWithLikeCount::getGroupId).collect(Collectors.toList())).stream().collect(Collectors.toUnmodifiableSet());
+
+        Set<Long> pendingDateGroupIds = groupDateRequestRepository.findAllByFromGroup(group).stream().map(GroupDateRequest::getToGroup).map(Group::getId).collect(Collectors.toUnmodifiableSet());
+        Set<Long> myLikeGroupIds = groupLikeToDateRepository.findAllByFromGroupMember(memberRecordOfUser).stream().map(GroupLikeToDate::getToGroup).map(Group::getId).collect(Collectors.toUnmodifiableSet());
+
+        List<DateableGroupResponse> likedDateableGroupResponses = likedGroups.stream()
+                .map(likedGroup -> {
+                    DateableGroupResponse response = DateableGroupResponse.from(likedGroup, RequestStatus.EMPTY, null, groupIdWithLikeCountMap.get(likedGroup.getId()));
+
+                    if (pendingDateGroupIds.contains(likedGroup.getId())) {
+                        response.setRequestStatus(RequestStatus.PENDING);
+                    }
+
+                    if (myLikeGroupIds.contains(likedGroup.getId())) {
+                        response.setLikeStatus(LikeStatus.LIKED);
+                    } else {
+                        response.setLikeStatus(LikeStatus.NOT_LIKED);
+                    }
+
+                    return response;
+                }).collect(Collectors.toList());
+
+        return new PageImpl<>(likedDateableGroupResponses, pageable, idAndLikeCountOfGroupsLikeToDate.getTotalElements());
     }
 
     @Override
